Order-product 테스트 완료

+++ order/order/src/main/java/com/luckylogistics/order/application/service/OrderService.java
@@ -3,19 +3,11 @@ package com.luckylogistics.order.application.service;
 import java.util.List;
 import java.util.UUID;
 
+import com.luckylogistics.order.application.dto.*;
+import com.luckylogistics.order.domain.entity.OrderStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.luckylogistics.order.application.dto.CompanyHubResponse;
-import com.luckylogistics.order.application.dto.DeliveryCreateResponse;
-import com.luckylogistics.order.application.dto.HubManagerEmailResponse;
-import com.luckylogistics.order.application.dto.MinusRequest;
-import com.luckylogistics.order.application.dto.OrderRequest;
-import com.luckylogistics.order.application.dto.OrderResponse;
-import com.luckylogistics.order.application.dto.OrderUpdateRequest;
-import com.luckylogistics.order.application.dto.OrderUpdateResponse;
-import com.luckylogistics.order.application.dto.ProductResponse;
-import com.luckylogistics.order.application.dto.UserResponse;
 import com.luckylogistics.order.application.event.OrderKafkaEventPublisher;
 import com.luckylogistics.order.application.external.CompanyService;
 import com.luckylogistics.order.application.external.DeliveryService;
@@ -70,6 +62,7 @@ public class OrderService {
 		// 배송 생성
 		// TODO: 배송 경로에서 포함된 ID를 통해 허브 서비스에서 허브의 이름을 받아와야함
 		DeliveryCreateResponse deliveryCreateResponse = createDelivery(order, productResponse, companyHubResponse, userResponse);
+        order.updateDeliveryToOrder(order.getOrderId(), deliveryCreateResponse.deliveryId());
 
 		// 출발 허브 관리자 슬랙 이메일 조회
 		HubManagerEmailResponse hubManagerEmailResponse = getHubManagerEmail(companyHubResponse.arrivalHubId());
@@ -128,18 +121,18 @@ public class OrderService {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow( () -> new BusinessException(ExceptionCode.ORDER_ID_ERROR));
 
-        //주문 변경 시 재고에 변경을 감지해야 함
-//        int before = order.getQuantity();
-//        int after = orderUpdateRequest.Quantity();
-//        int difference = after - before;
-//        //수량이 증가할 경우 그 만큼 재고를 감소시키고, 아니라면 추가한다
-//        if (difference > 0) {
-//            productService.minusProduct(order.getProductId(),new MinusRequest(difference));
-//        } else if (difference < 0) {
-//            //절댓값 사용해야 함 (0보다 작으면 빠지니까)
-//            productService.plusProduct(order.getProductId(),new PlusRequest(Math.abs(difference)));
-//        }
-        //order.update(after,orderUpdateRequest.request());
+        //주문 변경 시 재고의 변경을 감지해야 함
+        int before = order.getQuantity();
+        int after = orderUpdateRequest.Quantity();
+        int difference = after - before;
+        //수량이 증가할 경우 그 만큼 재고를 감소시키고, 아니라면 추가한다
+        if (difference > 0) {
+            productService.minusProduct(order.getProductId(),new MinusRequest(difference));
+        } else if (difference < 0) {
+            //절댓값 사용해야 함 (0보다 작으면 빠지니까)
+            productService.plusProduct(order.getProductId(),new PlusRequest(Math.abs(difference)));
+        }
+        order.update(after,orderUpdateRequest.request());
         order.update(order.getQuantity(), orderUpdateRequest.request());
 
         return new OrderUpdateResponse(
@@ -149,6 +142,12 @@ public class OrderService {
         );
 
     }
+
+    //FeignClient용
+    public Boolean checkOrder(UUID orderId) {
+        return orderRepository.findById(orderId).isPresent();
+    }
+
     //전체 조회
     public List<OrderResponse> getAllOrders() {
         List<Order> orders = orderRepository.findAll();
@@ -166,15 +165,15 @@ public class OrderService {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow( () -> new BusinessException(ExceptionCode.ORDER_ID_ERROR));
 
-//        if (order.getStatus() == OrderStatus.DELETED) {
-//            throw new BusinessException(ExceptionCode.ORDER_ALREADY_DELETED);
-//        }
-//        else {
-//            //가진만큼 재고에 더해야함.
-//            productService.plusProduct(order.getProductId(),new PlusRequest(order.getQuantity()));
-//            //이후 삭제
-//            order.delete();
-//        }
+        if (order.getStatus() == OrderStatus.DELETED) {
+            throw new BusinessException(ExceptionCode.ORDER_ALREADY_DELETED);
+        }
+        else {
+            //가진만큼 재고에 더해야함.
+            productService.plusProduct(order.getProductId(),new PlusRequest(order.getQuantity()));
+            //이후 삭제
+            order.delete();
+        }
         order.delete();
     }
 
@@ -184,14 +183,14 @@ public class OrderService {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow( () -> new BusinessException(ExceptionCode.ORDER_ID_ERROR));
 
-//        if(order.getStatus() == OrderStatus.ORDERED) {
-//            throw new BusinessException(ExceptionCode.ORDER_ALREADY_EXIST);
-//        }
-//        //롤백하면 재고 차감해야 함
-//        else {
-//            productService.minusProduct(order.getProductId(),new MinusRequest(order.getQuantity()));
-//            order.rollbackDelete();
-//        }
+        if(order.getStatus() == OrderStatus.ORDERED) {
+            throw new BusinessException(ExceptionCode.ORDER_ALREADY_EXIST);
+        }
+        //롤백하면 재고 차감해야 함
+        else {
+            productService.minusProduct(order.getProductId(),new MinusRequest(order.getQuantity()));
+            order.rollbackDelete();
+        }
         order.rollbackDelete();
     }
 }
